Leave balloted vers of spec, create STU1.md
We need to leave the May2019ballot version of the FHIRcast specification accessible  for historical reasons, and as we version the spec (pre-STU, STU1, STU2, etc), we want these versions to be manageable. We could use the github "release" functionality to look back in time at the previous versions, but that seems complicated for publishing them all together.