make preservesFhirId required?
From the SHALL phrasing in ReceiveEventNotification, we're saying setting preservesFhirId=`true` is required if you are preserving the FHIR ID. Should we go ahead and make the parameter Required? If not, should we add narrative somewhere to suggest that `preservesFhirId` can be treated as False by default?

(maybe something that can be left to the implementer, too)